--- a/additional.d.ts
+++ b/additional.d.ts
@@ -14,3 +14,9 @@ declare module '*.webm' {
     const src: string
     export default src
 }
+
+declare module '@mapbox/rehype-prism' {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const src: Pluggable<any>
+    export default src
+}