--- a/package.json
+++ b/package.json
@@ -44,6 +44,7 @@
     }
   },
   "dependencies": {
+    "@mapbox/rehype-prism": "^0.8.0",
     "@visx/grid": "^2.6.0",
     "@visx/group": "^2.1.0",
     "@visx/responsive": "^2.8.0",
@@ -53,14 +54,13 @@
     "d3-format": "^3.1.0",
     "d3-time-format": "^4.1.0",
     "gray-matter": "^4.0.3",
-    "highlight.js": "^11.5.1",
     "lodash": "^4.17.21",
     "next": "^12.1.6",
     "next-mdx-remote": "^4.0.3",
+    "prism-themes": "^1.9.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "rehype-autolink-headings": "^6.1.1",
-    "rehype-highlight": "^5.0.2",
     "rehype-slug": "^5.0.1",
     "remark-gfm": "^3.0.1",
     "rxjs": "^7.5.5",