--- a/src/lib/loadMarkdown.ts
+++ b/src/lib/loadMarkdown.ts
@@ -1,10 +1,10 @@
 import fs from 'fs/promises'
 
+import rehypePrism from '@mapbox/rehype-prism'
 import matter from 'gray-matter'
 import { MDXRemoteSerializeResult } from 'next-mdx-remote'
 import { serialize } from 'next-mdx-remote/serialize'
 import rehypeAutolinkHeadings from 'rehype-autolink-headings'
-import rehypeHighlight from 'rehype-highlight'
 import rehypeSlug from 'rehype-slug'
 import gfm from 'remark-gfm'
 
@@ -30,7 +30,7 @@ export const serializeMdxSource = async (markdownContent: string): Promise<MDXRe
             remarkPlugins: [gfm],
             rehypePlugins: [
                 // Adds syntax highlighting
-                rehypeHighlight,
+                rehypePrism,
 
                 // Adds "slug" IDs to each heading, for links and table of contents.
                 rehypeSlug,