--- a/src/pages/_document.tsx
+++ b/src/pages/_document.tsx
@@ -29,7 +29,7 @@ export default class MyDocument extends Document {
                     <link rel="preconnect" href="https://fonts.googleapis.com" />
                     <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="anonymous" />
                     <link
-                        href="https://fonts.googleapis.com/css2?family=IBM+Plex+Mono:wght@400;500&family=IBM+Plex+Sans:wght@400;600;700&family=Open+Sans:wght@300;400;600;700&family=PT+Sans:wght@400;700&display=swap"
+                        href="https://fonts.googleapis.com/css2?family=Source+Sans+Pro:wght@300;400;600;700&family=Source+Code+Pro&display=swap"
                         rel="stylesheet"
                     />