--- a/src/styles/components/blog/_LinkPost.scss
+++ b/src/styles/components/blog/_LinkPost.scss
@@ -7,11 +7,5 @@
         p:nth-child(2) {
             display: inline;
         }
-        code {
-            background-color: $light;
-            border: solid 1px $border-color;
-            padding: 2px;
-            margin: -2px 0;
-        }
     }
 }