--- a/src/styles/components/blog/_ReleasePost.scss
+++ b/src/styles/components/blog/_ReleasePost.scss
@@ -43,13 +43,6 @@
             background-color: var(--light);
             text-decoration: none;
         }
-
-        code {
-            background: var(--light) !important;
-            color: var(--dark) !important;
-            border: solid 1px var(--border-color) !important;
-            font-size: 85% !important;
-        }
     }
 
     &__help {