--- a/src/styles/globals.scss
+++ b/src/styles/globals.scss
@@ -3,8 +3,6 @@
 
 @use 'sass:math';
 
-@import 'node_modules/highlight.js/scss/atom-one-light.scss';
-
 @import 'node_modules/bootstrap/scss/functions';
 @import 'node_modules/bootstrap/scss/mixins';