--- a/src/styles/templates/_PostTemplate.scss
+++ b/src/styles/templates/_PostTemplate.scss
@@ -21,7 +21,6 @@
         ul,
         ol,
         pre,
-        code,
         figure,
         img,
         object,