--- a/src/styles/templates/_PostTemplate.scss
+++ b/src/styles/templates/_PostTemplate.scss
@@ -25,7 +25,8 @@
         img,
         object,
         table,
-        blockquote {
+        blockquote,
+        video {
             max-width: 100%;
             width: 40rem;
             margin-left: auto;