@bretthayes 2022-06-28
Awesome, thanks for pushing up that change @st0nebraker! Doing a follow up audit for all code blocks that are missing languages would be great! These are all of Prism's [supported languages](https://prismjs.com/#supported-languages).

For a little bit of backstory, I actually tried out rehype-highlight with highlight.js as seen in [my first commit](https://github.com/sourcegraph/about/pull/5520/commits/668704ec30c0a54ca6c39d41afac5e217b8b387a#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519R56-R63), but there were less supported languages in the core so I opted out and went for Prism. Prism has [239~ supported languages](https://prismjs.com/#supported-languages:~:text=This%20is%20the%20list%20of%20all%20293%20languages) versus [~40 in highlight.js](https://github.com/highlightjs/highlight.js/blob/main/SUPPORTED_LANGUAGES.md).

As for our font-size, the computed font-size is actually the same for code blocks and body text (16px / 1rem). However, it appears slightly different because of the font-family that code blocks use. I just did an audit of our fonts included on the site since we have 4 different fonts with different weights. I just removed some unused font families so that should increase our performance/load times a bit. I included a matching monospace font to compliment our base font so that should resolve the readability issue. 🤓 

Also just added a few bug fixes I noticed in [this commit](https://github.com/sourcegraph/about/pull/5520/commits/fb0a31b38515b9498b317bb1465ff4885d977aad). The code insights page video wasn't aligned and I noticed some video elements on blog posts that weren't within the container width constraint.