@bretthayes 2022-06-29
Just noting here for historical purposes; we gained a 6% overall performance boost by fixing our font families in commit [3cba42d](https://github.com/sourcegraph/about/pull/5520/commits/3cba42ddaf16a28e1e87272c389f3cd1008900e5). Doing a Lighthouse audit just before this PR was merged and after this PR was merged resulted in the following results:

- overall we went from 54% to 60% (6% difference)
- FCP went from 2.3 to 2s
- LCP went from 3.6s to 2.1s, and;
- CLS went from 0.014 to 0.013

![image](https://user-images.githubusercontent.com/1733936/176461320-92218086-5f26-4772-8088-4ec1a3413387.png)